+++ trace_file.hpp
@@ -67,7 +67,7 @@ public:
     void close();
     void flush(void);
     int getc();
-    bool skip(unsigned length);
+    bool skip(size_t length);
 
     virtual bool supportsOffsets() const = 0;
     virtual File::Offset currentOffset();
@@ -79,7 +79,7 @@ protected:
     virtual int rawGetc() = 0;
     virtual void rawClose() = 0;
     virtual void rawFlush() = 0;
-    virtual bool rawSkip(unsigned length) = 0;
+    virtual bool rawSkip(size_t length) = 0;
 
 protected:
     std::string m_filename;
@@ -152,7 +152,7 @@ inline int File::getc()
     return rawGetc();
 }
 
-inline bool File::skip(unsigned length)
+inline bool File::skip(size_t length)
 {
     if (!m_isOpened || m_mode != File::Read) {
         return false;
@@ -175,7 +175,7 @@ protected:
     virtual int rawGetc();
     virtual void rawClose();
     virtual void rawFlush();
-    virtual bool rawSkip(unsigned length);
+    virtual bool rawSkip(size_t length);
 private:
     void *m_gzFile;
 };
